refactor(api): move route /api/token-from-external-user to IAM context
