+++ api/src/identity-access-management/application/saml/saml.route.js
@@ -1,3 +1,6 @@
+import Joi from 'joi';
+
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { samlController } from './saml.controller.js';
 
 export const samlRoutes = [
@@ -25,6 +28,33 @@ export const samlRoutes = [
     config: {
       auth: false,
       handler: samlController.assert,
+      tags: ['api', 'authentication', 'saml'],
+    },
+  },
+  {
+    method: 'POST',
+    path: '/api/token-from-external-user',
+    config: {
+      auth: false,
+      validate: {
+        payload: Joi.object({
+          data: {
+            type: Joi.string().valid('external-user-authentication-requests').required(),
+            attributes: {
+              username: Joi.string().required(),
+              password: Joi.string().required(),
+              'external-user-token': Joi.string().required(),
+              'expected-user-id': Joi.number().positive().required(),
+            },
+          },
+        }),
+      },
+      pre: [{ method: securityPreHandlers.checkIfUserIsBlocked }],
+      handler: samlController.authenticateForSaml,
+      notes: [
+        '- Cette route permet d’authentifier un utilisateur Pix provenant de la double mire GAR.\n' +
+          '- Elle renvoie un objet contenant un access token.',
+      ],
       tags: ['api', 'authentication', 'saml'],
     },
   },
